Moving parallel stuff into tasks.py; WIP
